V 0.5: fixed bugs with labels

+++ stack.h
@@ -1,9 +1,9 @@
 #define STACKCONSTRUCT( _stack ) stack_construct(_stack, #_stack)
 #define STACKDUMP( _stack, err ) stack_dump(_stack, err, __LINE__, __FILE__)
 
-typedef double stack_type;
-#define OUT_TYPE "%.0lf"
-#define IN_TYPE "%lf"
+typedef float stack_type;
+#define OUT_TYPE "%0.2f"
+#define IN_TYPE "%f"
 #define TRANSLATE_FUNC atof
 
 const int          CANARYCONST1     = 0xFAAAAAAC; //First canary in struct stack_t value
